fix(agents): widen the prune cutoff before subtracting (java:S2184)
ToolResultPruner computed messages.size() - protectRecent in int before
Math.clamp widened it. chat.pruneToolResultsProtectRecent has no write-time
validation, so a large negative value wrapped the difference negative and
clamped the cutoff to 0, pruning nothing instead of everything before the
last user message. Casting the size to long uses Math.clamp(long, int, int)
as intended; results for any non-negative protectRecent are unchanged.

Adds a regression case, verified red without the cast and green with it.
./jclaw.sh test passes: backend 555 classes, frontend 1601 tests, stylelint,
lint, typecheck.

+++ app/agents/ToolResultPruner.java
@@ -47,7 +47,7 @@ public final class ToolResultPruner {
      */
     public static List<ChatMessage> prune(List<ChatMessage> messages, int minChars, int protectRecent,
                                           @Nullable Agent agent, @Nullable Conversation conversation) {
-        int cutoff = Math.clamp(messages.size() - protectRecent, 0, lastUserIndex(messages));
+        int cutoff = Math.clamp((long) messages.size() - protectRecent, 0, lastUserIndex(messages));
         List<ChatMessage> out = null;
         int stubbed = 0;
         long elided = 0;
